Functional WORKING - Changed alive time to 5 seconds because of slow threads (1 sec min) - Added Functional automatic timed functions - Removed obsolete code (not useful stuff) - Added catch for std::exceptions for further debugging on each thread - Finally finished thr_3 (Advanced Functional Thread) - Changed Block & Text data to smart pointers - Removed CUSTOM_EVENT_EXTERNAL_EXIT_CALL (Shared) because it's useless (obsolete) - Sprite_Base's 'twinUpAttributes' now works with std::shared_ptr of its data, not from a Sprite_Base itself - Removed duplicated code on SuperPair - Added size() and erase() to SuperMap - Renamed MILI to MILLI because it's MILLIseconds, not MILI haha - Updated main to test new functional feature and Button feature

+++ LSW5/Library/LSW5/hash/sha256.cpp
@@ -127,7 +127,7 @@ namespace LSW {
 
                 SHA256 ctx = SHA256();
                 ctx.init();
-                ctx.update(conv.c_str(), conv.length());
+                ctx.update(conv.c_str(), static_cast<unsigned int>(conv.length()));
                 ctx.final(digest);
 
                 char buf[2 * SHA256::DIGEST_SIZE + 1];
